Remove entry_dep
